# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple Rental Service server.",
        "version": "0.0.5",
        "title": "Swagger Rental Service",
        "contact": {
            "name": "bolszow@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        }
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get user list",
                "description": "",
                "operationId": "getUserList",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "username",
                        "description": "The name that to be fetched",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "description": "",
                "operationId": "createUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Create user object",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/users/{userId}/movies": {
            "get": {
                "tags": [
                    "user movies"
                ],
                "summary": "List of user movies",
                "description": "",
                "operationId": "userMovies",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "description": "User ID whose list of movies should be displayed",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user movies"
                    }
                }
            },
            "put": {
                "tags": [
                    "user movies"
                ],
                "summary": "Update list of user movies",
                "description": "",
                "operationId": "updateUserMovies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "description": "User ID whose list of movies should be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "New user movies list",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        }
    },
    "definitions": {
        "Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "movie": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Category"
                    }
                },
                "title": {
                    "type": "string"
                },
                "posterUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "user": {
            "type": "object",
            "required": [
                "username",
                "firstName",
                "lastName"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/movie"
                    }
                }
            }
        }
    }
}